Fix unread notifications
Reading one thread would mark all messages read. The expression
read=0 OR seen=0 was getting combined with the thread_id in the
provider like this: read=0 OR seen=0 AND thread_id=N, which was evaluated
as read=0 OR (seen=0 AND thread_id=N). The expression needs parens
so it will get evaluated correctly. Bug 2615680

Change-Id: Ib495a74e41346b835a576a8d6d53e809b839f93d