Deprecated language codes leaked in HTTP_ACCEPT_LANGUAGE header
Some old language codes, such as Hebrew (iw), were inserted in the
HTTP_ACCEPT_LANGUAGE header. Fix the problem using the same strategy
as in the browser code: replace deprecated language codes with the
new ones. Added a unit test that tests for standard behavior and tests
one of the deprecated languages. Bug 2600706.

Change-Id: I5ac29618c9c7f5cbbafdb24c43367c983647eaa8