Turn off smooth scrollbar
The dev docs say, "should be set to false if your adapter will display
items of varying heights." Since ComposeMessageActivity shows items of
varying heights, this was causing the scrollbar to grow and shrink while
scrolling. With this change, the scrollbar doesn't scroll smoothly, but
the thumb now accurately reflects the scroll position. Bug 2709895.

Change-Id: I1130a1001b093adc11b85c38c095fdc3792b4c1a