Fix mismash message
DO NOT MERGE
A particular set of repo steps involving creating a draft, then deleting
the draft and then receiving a new message can lead to a situation where
a new incoming message gets the old thread id of the discarded
draft. The compose activity can end up displaying the recipients of the
old message with the contents of the new message. Recognize this dangerous
situation and bail out to the ConversationList where the user can
enter the compose activity in a clean manner.
Bug 2793874

Change-Id: I04727de3908e3c50d4687cb6e2a9cc6589daed9c