Mms: Delivery report should display recent message details.
On sending SMS to multiple recipients with delivery reports enabled,
the delivery reports for all the recipients display the details of first
recipient since the time stamps are same. To fix this, need to query in
the ascending order and display the details of last message which is the
most recent message.

Change-Id: I9917a7a3fd56daceccdc9fdc5f3cf1294c22c320