Handle OutOfMemory when decoding images
Decoding images and videos in the messaging app appears to be a
vulnerable point for running out of memory. Catch those exceptions,
log an error, and continue on without loading the image. Most of these
"catches" are around situations where there's been a problem decoding
the original image in the first place, and now we're simply trying
to decode a small missing stand-in image. However, since BitmapFactory.decodeResource
seems to be in an inordinate number of bug reports, I want to log every
use in the code in case of failure.
Handles the 2nd part of bug 2994057.

Change-Id: Iaa8035f6ef28385b353d31507b9753934881d3e8