Update patch set 1
Patch Set 1: (2 inline comments)

The old code was fragile and incorrect; this is definitely an improvement. It looks correct and should improve the app.

But it isn't as simple as it could be. Since your new chunked case will also work for the non-chunked case, you can drop the if/else altogether and handle both with your new code. And there's a few more small simplifications still on the table.

Overall a good patch.

Patch-set: 1
Reviewer: Jesse Wilson <1002348@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Verified=0