Set local time as timestamp when MMS has been downloaded
Previously, the timestamp of notifications was the current local time,
but as soon as the MMS was downloaded it got the timestamp from the
PDU. This could cause problems if timezones weren't set correctly etc.
Now a downloaded MMS will always get the current local time as its
timestamp, regardless of when it was first received.