Extended search for duplicate MMS messages during download
MMS messages sent from some test servers all have the same message ID,
which results in the message being deleted from the Inbox after a
first message has been downloaded. This is fixed by requiring that the
IDs and subjects of two messages are identical to regard them as
duplicates.