FIX: CM issue 1733 / AOSP issue 2989
This is a first idea to fix CM issue 1733 / AOSP issue 2989.
The original implementation just stored the local time as the
sent time and not the time that's indicated in the SMS message
by the SMSC. I intended to add a second Inbox.DATE value for the
real SMSC time but that would need a lot of changes on several
places in the source.
So this request can also be seen as some kind of RFC.

Change-Id: I5d115a8a32f23bad858324c60115a7f32bbfb60d