Set the default MMS max message size to 307200
The default max message size needs to be set to 307200 to handle
EU's typically smaller message size. This change was originally made
to a GB release build and made its way to HC and master, but didn't
quite make it to GB itself. The overlays for AT&T, TMO, and others
override this default as necessary to support the limits of their
systems. Bug 4124807

Change-Id: Ia8d04ca873f9c91ae47ac5ef860553befb8ae43b