Update patch set 1
Patch Set 1: Verified; I would prefer that you didn't submit this

The change merges and builds fine. However, I have some concerns about what it does, and I seem to recall that 300kB was the threshold for interop. I don't think that we should should the value that drastically across all devices, i.e. in my opinion this should be a configurable parameter that should still default to 300kB.

Patch-set: 1