Add new send sms intent
Add a new intent to send an sms message without any UI interaction.
Presently, the Messaging app can already respond to intents with
the android.intent.action.SENDTO action, but this brings up the
ComposeMessageActivity primed with a destination and message. It's up
to the user to review the message and hit send. This new intent will
send the message to the user without the user seeing the compose screen
and having to hit send. Apps could already send sms's by talking
directly to the SmsManager. Using this new intent, however, has the
benefit of adding the outgoing message to a thread so the user will
be able to see what was sent. In addition, sending a message via the
Messaging app compare to the SmsManager will add automatica retry when
there's no connection, etc. This new feature could be used by the phone
app so the user could easily respond to an incoming phone call by sending a canned sms that
reads, "I'm busy, but I'll call you right back." A caller of this new
intent must have android.permission.SEND_SMS_NO_CONFIRMATION. An
accompanying CL adds the private permission to the framework.
An included test app tests the feature. Bug 4563486

Change-Id: I9292ea6d52e135cdabcfe68091d4d24d15524c97
Signed-off-by: Tom Taylor <tomtaylor@google.com>