Message app: Inconsistent state
The messaging app died and was restarted by selecting a mms notification.
The app was restarted with a saveInstanceState which loaded conversation A.
Then the app was handed the new intent from the notification for
conversation B. Meanwhile, conversation A's query finished, but the threadId
no longer matched that of conversation B. When the code starts a new
query, it tries to cancel any pending queries, but the docs says that's
not a guarantee. We now detect this state and restart the query to match
the current conversation. Bug 5179301

Change-Id: I04df8dc90c2ee152d255ba42a1a699106e03a2a4