5198894 Remove saveInstanceState
ComposeMessageActivity has code in onSaveInstanceState to save
the working message, the recipients, and uri to the bundle.
In onCreate, this bundle was used to restore the conversation
and working message, regardless of the current intent. Since
the code is already saving and restoring the draft, let's
remove the saveInstanceState code. Bug 5198894

Change-Id: I038269020e4f02e1390b3a62a13278cfd72303ac