Failed to initialize the slide-list on selecting slide show attachment
The key to this bug is to add an empty subject and then add a slideshow.
Because the subject is empty and onStop on ComposeMessageActivity
is getting called when the slideshow editor is displayed, CMA is saving
the draft. Since the subject is now empty, the draft gets downgraded to
an sms and the mms draft gets deleted behind the slideshow editor's back.
The safest solution is to not be so eager to remove an empty subject.
For this edge case, an empty subject will be removed when the user hits
send and the message will be sent as an sms if possible. Bug 5181255

Change-Id: I9c08f744928211af0a336f67eb32b258f58f4622