Attached picture is not saved in drafts when a received message is opened from notification bar
Our mConversation.sameRecipient() code didn't work on uri's from
notifications that contained a threadId instead of a number. Therefore,
an incoming message never matched a new draft that had a threadId of
zero, even when the recipients matched. When we decided a conversation
matched, the code didn't update itself or the working message with the
updated conversation that has a valid threadId. Bug 5250265

Change-Id: I6c97d5c53cf75d97b652b7635020c2cb30855761