Fix a number of draft-related bugs
This CL contains a number of draft-related fixes. One is timing related.
While we're building the PDU for a draft, particularly a big PDU that
involves image resizing, the thread gets created first, before the PDU.
Meanwhile, ConversationList can finish its first query and delete any
empty threads. We guard against that several ways, but even still I
see cases where the thread gets deleted while the PDU is saving. When
that happens, pdu persister creates a new thread. Our working  message's
conversation would still point to the old thread id. With this change,
we get the possibly new conversation after updating the pdu. Note that
all of these cases occur when this is a draft or the first message
in a thread. Also, there's a better fix for the zombie draft bug. We were
initiating deleting drafts when we cleared the conversation between message
sends. This sometimes caused us to delete the draft we were in the process
of sending. Now draft deletion only occurs at the end of sending. There's
still a small timing window where the draft can get deleted. That'll be
fixed in an upcoming CL.
Bug 5340360. Bug 5343360. Bug 5299114. Bug 5353623. Bug 5354900

Change-Id: I0f661959aed38d59daf5fa27b1f006953e15b80a