Not able to fast forward the playback forward the playback for Audio files attached in MMS
While playing back an audio, tapping on the screen brings up a control
that looks like it should control the audio. Unfortunately, that control
is the control for the slideshow -- jumping to the next slide, pausing,
etc. When there's only a single slide, such as a sound attachment, we
now hide the slideshow controller to avoid the confusion. Bug 5360885

Change-Id: I69d6c2c65105b2b9fd6b25868359da74764f9e45