java.lang.IllegalStateException: save() called after discard()
Bug 5372845
How can the monkey find this? Well, it can click on the menu "Delete thread"
and then click on anything else that'll cause saveMms to be called before
the onDeleteComplete callback gets executed. We've had this intentionally
thrown exception in the code for years and we're confident that we're
not trying to save deleted messages. With this change, we now do what we
do in saveDraft and that is bail out when the working message has been
discarded.

Change-Id: I262b7c153add3ebbed1044f4c51fe6899b963cd8