Weird phone number stored in canonical_address table
Bug 5375782.
Very tricky bug. In the recipient editor, we always leave a trailing
comma to make it easy for users to add additional recipients. When a
user types (or chooses from the dropdown) a new contact Mms has never
seen before, the contact gets the correct trailing comma. But when the
contact gets added to the mms's contacts table, contacts sends out an
onUpdate to CMA. CMA would recompute the recipients and since the
recipient editor was still visible, call mRecipientsEditor.populate(recipients).
This would replace the recipient that had a comma with a recipient
without a comma. When a user manually added a new comma to add another
recipient, this would eliminate the span inside the text. The span contains the
number part of "Fred Flinstone <123-1231>". Hence, the whole
"Fred Flinstone <123-1231>" would be considered the number of
the first recipient and get entered into the canonical_addresses table.
The fix is very easy. All recipients have commas. There's also a change
to rebuild the contacts cache when we delete a thread. We do that because
the provider is now deleting stale contacts when threads are deleted.

Change-Id: I8fdde2ba64e934c9b539fa0fae433ef23fd254bc