SMS messages being delivered to wrong contact
Bug 5319828
Last ditch escape hatch at detecting an sms being sent to a wrong
contact. From what we've seen of a database where sending to the wrong
contact had taken place, the in-memory Conversation cache had the
correct threadId, but the wrong recipient. With this change, we do
two checks to verify that the recipients for a particular thread match
the recipients associated with the thread in the database.

Change-Id: I8f7c194243e6da093461c1aaae4119f05390582d