My photo isn't displayed next to messages that I sent
Bug 5361233. Bug 5226635. Bug 5383721.
Add the concept of a "Me" contact to our contacts. It has the correct
uri of the profile so fastrack works. Also, it will load the profile
regardless of the state of the user's phone number. Tapping on your
own quick contact badge will show the profile.

Change-Id: Ib6ea288c9596a001f167a151aaeec6fe2c7c8e2c