MMS Service and uaProfUrl config attribute late binding check
bug: 5373677
When the uaProfUrl config attribute is missing from mms_config.xml,
we want to allow SMS but crash MMS.  Therefore, MMS should crash as
late as possible.  We will warn with log messages if the uaProfUrl
is missing, and only crash with a runtime exception if the user
tries to send an MMS with an absent uaProfUrl.
Change-Id: I960e752538e3c3ee7d474327b89522c745a8a0c0