Messaging App does not discard draft after deleting it
Bug 5376774.
It was tempting to remove this block of code since the provider is no longer
automatically deleting empty threads. However, the comment is correct.
When the conversation has no messages, the user can still add recipients
and that would change the threadId. The solution is to delete the draft here.
We'll create a new one in onStop if necessary.

Change-Id: I3f79cb6c07ca67ba7280e2f78be9793581b87f84