Fix memory leaks in SMS/MMS list items: b/5393957
The main leak was from the fact that recycled and released views
were still being kept in mAddressToMessageListItems, and each view
could store references to a message item with attachments. It turns
out that that map is actually never used because no one calls
notifyImageLoaded. So I'm removing the map and the method.

Change-Id: Ie1fd787c9dee6b6d803d1e778a615665b2299bfd