bugfix 5342861 The search result highlighted result is not seen on the screen for MMS
Part 1 of 2 part change to fix this bug.  This changes the suggestions provider
to show word matches rather than whole result matches.  Clicking on a suggestion
then takes you to the normal search result page which lists specific messages.
Clicking on that takes you to the chat with the appropriate highlighting and
positioning.

Change-Id: Ic3f4a2c5609c165f30024d5697f4e52571abb565