NPE in DraftCache
Bug 5416424. Bug 5416250.
The cursor wasn't checked for null, but still closed in "finally".
This code is old, but now exposed because we rebuild the draft
cache when deleting all threads. I tried to repro the bug and stepped
through the debugger and we're simply getting back an empty (but
valid) cursor. Fixed the code to handle null cursor.

Change-Id: Iaac6dd8791cd12ad33e33fb9c03bee088d36198f