Don't save draft with no recipients
Bug 5421126.
If you create a new message and type something or attach something,
but don't add a recipient and then select the notification of a new
message, we'd save the draft of the message. The draft would be saved
with no recipients. It looks bad in ConversationList. Normally when
the user backs out of CMA in this state, they're prompted with a
warning that the draft will be discarded. In this bug, however, CMA
doesn't get the opportunity to wait while the user confirms the
tossing of the draft. The solution with this change is to not
save a draft with no recipients. Tested a previous bug in this area
5127453 to verify no regression of that bug.

Change-Id: I6be813e241921af49beb497092be8f2cacb23abb