Show progress spinner when loading MMS attachments
bug: 5354767
When attaching to an MMS, loading of the media is done on the UI thread.
In the case of Picasa photos, loading can take several seconds.  Therefore,
this fix loads attachments in the background while a progress dialog is
displayed on the UI.  The dialog blocks user input, and appears only after
.5 seconds.  The fix has also been applied to video and audio attachments.

Change-Id: I30ab8b4aa3c91b4c401ab89c17e72efb84860879