MMS attachments now load asynchronously while ui shows progress dialog
Originally a bug fix for MMS attachment progress spinner
Bug: 5473834

Loading attachments used to happen on the ui thread.
This was bad news bears for loading times > 1 second
(a specific case being Picasa photos).
Now loading attachments now happens in the background while
showing a blocking progress spinner in the ui.  The
spinner only if loading takes longer than .5 seconds.
This applies for Images and Video, both for Messenger
retreiving media as well as receiving an intent.

Change-Id: I09ac701a8cefe3dc0c3d0dc91b0a77933854872c