Device unable to send SMS message to nicknames which contains characters and numbers
Bug 5453853
If an address looks like an alias and aliases are enabled, then treat the
alpha-numeric address as an address, rather than stripping non-dialable
digits and turning the address into a number. I tested this change in
conjunction with addresses such as "T-Mobile" to verify that those
addresses still work.

Change-Id: Ibd1a60ffddb481c977196c76bf13c8ce0a0da912