Can't add smiley to subject
Bug 5453810. Bug 3382512
If you use the menu to add a smiley when the subject editor has the focus,
the smiley will be added to the message body instead. That's now fixed.
Smiley icons are now displayed in the ConversationList and in the
subject of messages. The nose-less smiley icon :) is now parsed as
"happy" like it is in Talk.

Change-Id: I2685fd7946881aceb89617cf54c4de0b5a17650e