Incoming slideshows with sound and video will crash!
Bug 5419450
While investigating 5419450, a bug with multiple sounds formats and
video formats, I sent them all to my phone via an email mms gateway.
On the phone's side, we tried to build a slideshow from the attachment.
The mms code doesn't allow a video and sound to be added to the same
slideshow. The UI enforces that. However, when we receive a slideshow
with mixed sounds and videos, the code would throw an exception. With
this change, the code will ignore parts that can't be handled.

Change-Id: I298ee1722afc46593c48a95ceb825ae9c15c41b3