Download attachments are not saved in correct folder on sdcard
Bug 5491310
When saving the content of a local slideshow (one the user sent rather
than one received), the location of videos is a temporary location
on the sdcard. Don't use that location as part of the destination
file, rather just use the leaf filename. With this change, saving
any mms (whether sent or received) saves all the parts in the standard
Download directory.

Change-Id: I3b993865298adbcd9518a446514ff687052aba81