MMS attachment over 1.2MB
Bug 5504281
Sometimes the Messaging app will attempt to send a mms message that is
over the size limit because the message picture attachments don't get resized.
This can happen when the code attempts to read the input stream's
length and gets back zero. A shortcut in resize sees that the len of
zero is less than the targeted size for the picture and slkips the
resize step.

Change-Id: I443ef9a537010b1c685f20760662398cfcfa8e48