Not showing error when message is too big
Bug 5504281
When a user adds a large video and lot of pictures, there's not enough
room to compress the pictures and the app shows the user an error
dialog. When the dialog is dismissed, the user sees the message in
the "Sending..." state (even though the message wasn't sent). The problem
was very simple, the code for setting the error in the pending_msgs table
was using the wrong key so the message wasn't marked with an error. With
this change, the user sees the message tagged with an error, they can
tap it to edit it, and remove some of the slides and resend it. Granted,
it's an edge case, but it also explains why unsent message were showing
up as "Sending...", rather than marked with an error. Also, adding a log
for message resizing, since I've needed this debugging code twice in the
last few days.

Change-Id: I9ca5d20e2e7e45953098f4098cc154b2855a4f44