PduBodyCache.getPduBody cache not cleared
Bug 5526870
The PduBodyCache was never getting cleared. If you deleted a message
and then added a message and they had the same id, we'd return the old
pdu of the deleted message. It turns out that the PduBodyCache is
redundant because the pdu is already getting cached at a lower level in
the PduCache. Make sure we delete the cache entries associated with
individual messages and threads when those messages and threads are
deleted. While stepping through the change in the debugger to verify
it, I found that the new temp file names that start with '.' stripped
the filename incorrectly.

Change-Id: Ice4e453470eb1186ea69ebe6eca750bb230c5205