Sometimes the Call icon doesn't show up
Bug 5545195
The Call icon is added to the menu in onPrepareOptionsMenu. That only
gets called when the activity is created and when the user taps on the
menu button. However, some of those menu items are shown on the action
bar so we have to dynamically update the menu items. Whenever we set
the conversation in the working message, update the menu bar.

Change-Id: I9fd09f851b17c05704e82d93a754541c1b3f6a00