Revert "5198894 Remove saveInstanceState"
This reverts commit 7d128943d1b6377fdbe399eb919970ab9544867e.

At the same time, it fixes bug 1556157, where a picture sent via mms
from the camera, will show up again in a draft when ComposeMessageActivity
is killed and restarted. That behavior was caused by the fact that when
an activity is re-created with an instance bundle, it's passed the
original intent to attach a picture. With this fix, we don't try and
reattach a photo. In addition, killing the activity and restarting will
restore the activity to its pre-killed state (i.e. preserve the message,
subject, etc).

Change-Id: I2059d289bce6f40f11c286436c2fa0971d0ff273