MMS delivery reports and read reports are displayed as sent/read while the message is still being sent
Bug 5533745
This is a regression from Eclair. We're showing icons that indicate a message
has been delivered and/or read, when that might not be the case. With this
change, we do what we did in Eclair: show an icon that indicates there
might be more information available under the details menu.

Change-Id: I56a7677b3e0e675eee8c0adcd53f89681d50a8a8