Show "Loading conversations" instead of "No conversations" while querying
Bug 5717044
The empty view is shown when the adapter has no items. The empty view,
which used to read, "No conversations", can show up during a query. With
this change, we follow email's model and show "Loading conversations". If
the result of the query is no items, the message is changed to "No conversations"

Change-Id: I8e95eb134dd0ba73c2c2c0b60dc3a178302f1504