Wrong message deleted after changing device orientation while deleting a message
Bug 5787017
When a longpress menu item was selected, the old code relied on the cursor
being the same as when the tap took place. When the device orientation is
changed, the adapter rebuilds the list and the cursor is left at a random
position. With this change, the code saves the tapped message item and
ultimately uses the actual item when/if a menu item is selected.

Change-Id: I2f5af609b2ac918f1ffe613fc8e5397afff5ce64