Implement back intent on notifications
Bug 5214808

When the user selects a sms notification while in another app, they're
taken to the message conversation. Now when they tap back, they'll go
back to the conversation list. Tapping back again will take them back to
the activity they came from. This is the new standard in ICS.

Change-Id: I3a3dd97f6c895edd64656abf2995d17474950d11