SMS number input should auto-format like the dialer does
Bug 5851956

The recipients editor will now auto-format the number as the user types.
This only works on the first recipient, which is going to handle the
majority of the use cases. Holding off for now making it work with
multiple recipients until we decide whether we're going to implement
the "chips" ui (see b/5066294) for selecting recipients. While this
new code only works for the first recipient, it doesn't affect adding
additional recipients -- that works the same way it always has.

Change-Id: I9552997721829a5a84f15e6b5a3438639de88602