Crash in MMS on discarding a Message
Bug 5899557

There are two serious bugs here.
The first bug is that UriImage.getResizedImageData would get stuck in an
infinite loop when the picture data couldn't be decoded. Normally this loop
handles the case where an exception is thrown because the image is too big,
but in the case where the data is empty or perhaps corrupted, BitmapFactory.decodeStream
will return null and the code will end up in an infinite loop. This is done
in MmsSender in a background thread, so it doesn't affect the UI, but the
user sees that the sent message isn't showing up in the message list.
The second bug is when an exception is thrown, CMA puts up a dialog.
In the repro steps of this bug, backing out and the coming back in caused
this exception to be handled after (or during) the activity was finished.
I was easily able to repro this by hitting send and back quickly.

Change-Id: I3fe780119f15ef7d99dae45a3ebd2b4b56688e85