Some pictures not showing up
Bug 6092116

The pdu loaded callback was only getting set if the slideshow
was non-null. It should be the other way around.
Since the slideshow gets built at the same
time as the pdu gets loaded, list items weren't getting told
when the pdu was finished loading. With this change, the
code registers for the pdu loaded callback when the slideshow
and pdu haven't been loaded yet.

Change-Id: I93773f2a1e79a45d8cf6940b4a7f8f13b232ed4f