Increase cache size and put downsized thumbnail images on disk
Bug 6093907

Use Gallery code to compress thumbnails and store them in the
app's cache dir (not on the sd card). Increase the size of the
in-memory thumbnail cache to 16 images. Set the pdu and image
loader threads to a lower priority than the UI thread. Don't
extract the empty image bitmap on the UI thread. In fact, don't
show the empty image bitmap until we're sure the image can't
be decoded. Don't set the image bitmap multiple times. Remove
excess color list background that slows things down. Clean up
the thumbnail loader and get rid of the width, height params.
Lower the pdu cache to 16 items to match the thumbnail cache limit.
Modify SimpleCache to support both hard and soft references and
use hard refs for the small pdu and thumbnail caches. The result
is much faster and less janky scrolling.

Change-Id: Ifc0d590d4ab27bb66d306b681082b4e95cf3d64d