[VzW Issue] MMS picture file attachments disappeared when sent to an exchange email account (Sam_Int_36081)
Bug 6116662

Use the original filename if available as the name of the attachment.
Prior to this change, the code was using the leaf name of the uri
handed to the app from the gallery. This was typically a timestamp.
This was never a problem, but apparently MMSs mailed to Exchange
is a problem because it can't deal with images without extensions.

Change-Id: I17cac0113439fece6986459925d72dfddf41528f