Scrolling through a Compose Message Activity message thread is not 60 fps.
Bug 5928605

We're up to 57 fps for a conversation with only mms picture
attachments (but only 2-3 listitems are on screen at once) and
up to 48 fps for a conversation with only sms items (8 listitems
on screen at the same time). This CL adds a couple of tweaks.
One, the date wasn't getting shown on mms items. Two, the
expensive message building wasn't getting cached.

Change-Id: Ibcf214e571c9bbb91ba78a48321445633f9d9abb