ANR on the Messaging application, when editing the attachment
Bug 5973755

When slideshows are involved, there's the possibility of an ANR because
the slideshow is saved on the UI thread. With this change, the heavy-duty-
disk-writing work is done in a background thread. If the task takes longer
than a half second, a modal progress dialog is displayed. The various
target activities, such as SlideEditorActivity and SlideshowEditActivity,
still read/save the PDU on the UI thread. They should use the PduLoaderManager.
A separate bug has been filed.

Change-Id: I384968d406fe10e192abf1614e3ce0b27f20c93d