Empty forward draft not removed
Bug 6222154

A number of problems to fix for this relatively innocent looking bug.
First, the PDU was getting persisted on the UI thread. Second, persisting
a PDU automatically creates a thread, but on the forwarded message side,
the thread id wasn't getting set. Since the thread id was zero, the
working message code didn't think there was any draft to delete.

Change-Id: I3021b864bdf9876eb8c66e4048295516b085406f