We should not auto convert multiple SMS to MMS messages
Bug 2216705

Restore the functionality of the original fix from:
    https://android-git.corp.google.com/g/#change,40101
but keep the behavior from the subsequent CL:
    https://android-git.corp.google.com/g/#/c/102637/

With this change, the messaging app's sms multipart functionality is
still controlled by the mms_config.xml flag "enableMultipartSMS". When
this is true, the user can type a message up to 2000 characters and it
will be broken up and sent as parts. If enableMultipartSMS is false, as
soon as the user types more than a single segment (i.e. 140 chars), the
message will be sent as mms. Like the original fix, there's no part
threshold anymore that determines when a long message will be converted
from a multi-part sms to an mms.

Change-Id: I5a70f27876417786a2d79436614cb02cc8c69f82