Sending and receiving MMS over 4G is flaky
Bug 6094676

On Sprint we were getting back an empty, but not null APN name (from Sprint)
during the transaction handshake. Then the code would try to look up that
empty name in Telephony's carrier table. It would fail. With this change,
MMS send/recv is restored on Sprint (and perhaps other carriers).

Change-Id: Id5b47e513a72c7e9758e4fc1acd12ad03c213169