Red chip is shown for the number though the number is correct
Bugs 6391820, 6373342, 6387440, 6404611

Add a Validator to mms's recipient editor. Without a Validator, the chips
code will always construct a chip and color it red, indicating it is
invalid. Since mms has to handle email addresses, phone numbers, short codes,
and aliases, anything is valid. With this change the chips are created with
a gray background to match that of gmail (b/6391820). Also, supply a custom
chips recipient list item so when we list the phone numbers, we can display
the type as well (i.e. Mobile, Home, etc.). Also fixes coming into the
Messaging app via an intent (i.e. from Contacts) and showing the chip for
that person, instead of just the number.

Change-Id: I6826ecbfbd2e389b7d3fa5852059807db9431ac1