Messaging app crashed
Bug 6350665

There's a race condition in Telephony.Threads.getOrCreateThreadId when it
calls MmsSmsProvider.getThreadId. If a user deletes a thread and a message
for that thread arrives at the same time, it's possible for the delete
to remove the thread found in getThreadId. With this change, getOrCreateThreadId
is wrapped by a function that blocks if we're doing a delete. I tested this
technique by adding a 10 second sleep in onDeleteComplete. The SmsReceiverService,
which calls getOrCreateThreadId is fine being blocked, since it does all of
its work on a background thread.

Change-Id: Ia9fc04052adac15c2252034006dd6cdd5c02f108