Random person showed up as the sender of an email originated sms
Bug 6132498

When loading the contact for an email address, the correct person id
was getting overwritten by the email id. In many or most cases, this
would be an unknown contact id, but in this particular bug, happened
to be the contact id for a person in the contacts db. With this change,
the correct contact info and avatar are fetched for an email address.

Change-Id: Iec03a892bcfe1408a7ab68972c0794eb86c88691