Jank when sending a picture
Bug 6492075

The default when asking what kind of attachment we've got returned TEXT,
rather than unknown. Depending on when this function is called, the input
param might be null. In any case, when we don't know, the function needs
to return ATTACHMENT_TYPE_NOT_LOADED, not TEXT. When TEXT is returned,
MessageListItem hides the mms view. When the pdu/slideshow is loaded,
then the mms view is shown again, causing the jank. With this change,
the mms view remains visible throughout the sending process when there's
a picture/sound/video/slideshow attachment and the user sees no jank.

Change-Id: Iecd6067b164ee00a7e21c7ef753d6920188bda14