Fix a DEAD-LOOP in updating the recipients
It will be in a dead-loop between updateContact and
its callback method onUpdate() in ComposeMessageActivity
if we didn't wait the quering of contacts.

Also we don't need to reset the number which read from
RecipientEditor to the number of the contacts queried,
since it will be set a wrong number ends with the special
characters such as '*' to the contacts which has queried
and already be saved in the phonebook.

Change-Id: If89e4af54db38ad06e8d7742c549fcdd25ccb661
Signed-off-by: Xingrun Chen <cxr514033970@gmail.com>