Fix bug on copying/forwarding/viewing details of SMS message
When viewing an SMS conversation, sending a new message and
immediately long-pressing on that new message and immediately
select Copy, Forward, or View Details, on the context menu,
the app will act as if the message selected were another message.

Also reported by several people:
http://code.google.com/p/android/issues/detail?id=28009
http://code.google.com/p/android/issues/detail?id=26827
http://code.google.com/p/android/issues/detail?id=25569

This is caused by the app thinking that the
message selected is the same as the message pointed by
the position of the conversation's Cursor.
Actually, the position of the Cursor has been modified by
the ListView adapter calling getItem that sets the Cursor
position to the position of the last getItem, occuring
when the list refreshes itself due to the incoming new message
(which is the message that has just been sent).

This commit fixes the bug by:
- (for other than View Details): Passing the message item
itself to the handler of the context menu, so we're sure
that we're not copying/forwarding/deleting another message
- (for View Details): Creating another cursor whose
data is the same as conversation's cursor, and setting the
position to be the same as the selected message's position.

Change-Id: I2d9c412e6fa3683bc4f0c1710aba016daa68ccdc
Signed-off-by: Yuku on yuku4 <yukuku@gmail.com>