Update patch set 1
Patch Set 1: I would prefer that you didn't submit this

Hi Roger, great bug! I tried your fix and found another bug (already existing -- you didn't introduce it). I always test by sending messages to myself. After I delete all the messages and type a new one and hit send, it will send, but won't show up in the current thread. Instead, a new thread will be created and my sent message will not show up in any thread. I tracked down the root of the problem and it's because when all the messages from a thread are deleted, the thread gets deleted by a trigger in the database. ComposeMessageActivity doesn't know this and its mConversation still points to a non-existant threadId. I put in a fix in onQueryComplete to fix this and it fixes both your bug and my bug:

                    // Adjust the conversation's message count to match reality. The
                    // conversation's message count is eventually used in
                    // WorkingMessage.clearConversation to determine whether to delete
                    // the conversation or not.
                    int cnt = mMsgListAdapter.getCount();
                    mConversation.setMessageCount(cnt);
                    if (cnt == 0 && mConversation.getThreadId() > 0) {
                        // We just deleted the last message and the thread will get deleted
                        // by a trigger in the database. Clear the threadId so next time we
                        // need the threadId a new thread will get created.
                        mConversation.clearThreadId();
                    }

Patch-set: 1
Label: Code-Review=-1