Deleting individual sms message didn't work
Bug 6538916

Also fix a bug reported by an external developer where individually
deleting all the messages in a thread, then typing a message (w/o sending it)
and then going back to the conversation list shows "Saving draft" but
no draft or thread gets saved. In addition, I found that if you individually
delete all messages, then type a message and hit send, the thread disappears.
Both cases are due to the fact that when all messages are deleted from
a thread, the thread gets deleted. We need to recognize that situation
and clear the conversation's threadId so the next time the threadId is
needed, a new thread will be created.

Change-Id: Ic3aa8c3bf61709a75b5dd9d172c46fdef73d342b