Update patch set 1
Patch Set 1: I would prefer that you didn't submit this

I fixed this internally in the jellybean branch and the change will trickle down to master. You can abandon this CL. Thanks for finding the bug and proposing a fix! As I mentioned before, I fixed this a slightly different way that fixed the root of the problem (the thread was getting deleted).

Patch-set: 1
Label: Code-Review=-1