Fix "The forwarded msg cannot be discarded"
The forwarded msg is different from the msg which
sent from 3rd-party, it is a draft with the content
"Mms.Draft.CONTENT_URI", but the draft state wasn't
set anywhere. So we should force to set the state
after the conversation of workingmessage was set.
Otherwise, it cannot be discarded since mHasMmsDraft
was false.

Change-Id: I52e023da61b29701a2c3147812e82c3be159cefe
Signed-off-by: Roger Chen <cxr514033970@gmail.com>