Backing out of forwarding an mms leaves a row with no thread_id in the pdu table
Bug 6600669

This bug can also be reproduced by creating an mms message and then backing
out. You'll get a dialog that the message has no valid recipients and will
be deleted, but because the thread doesn't have a thread id (it has
no recipients), the pdu will get orphaned in the pdu table, never to be
deleted. With this change, when deleting a draft without a thread id, the
orphaned pdu(s) is delete as well. While I tested this with the various
combinations of mms attachments, forwarding, sending normally, and drafts,
there's some risk with this change because of the complexity of mms. In
the future, the code should be examined and changed so that a pdu with
no thread id is ever entered into the DB, but that's a much larger change
and research project than can be dealt with right now.

Change-Id: Ie0641873f052c9551fcd7ba56058cdb97287220a