Unable to save picture attachment from received MMS message
Bug 6615046

I was able to repro this by sending a picture attachment to my Sprint
phone via Sprint's gateway 12313123123@pm.sprint.com. The received SMIL
contained the picture, but also contained a text attachment and saving
that was causing the ContentResolver to choke on the type. The fix is
to skip the non-picture, video, and sound parts. This broke in JB
implementing the new drm. The fix is to put in a check that was in the
code for ICS.

Change-Id: I9ce5dcd7bac322b398d11f1d22cf59bc4755fcf0