bug #6664404: fix gratuitous scrolling in MMS compose activity.
- while on Compose activity, with the Message list view scrolled up somewhere, if the screen powers
off and back on, CMA requeries the message history, then calls to scroll to the bottom of the
message history. This behavior was on ICS and prior. However, if the IME keyboard is up while the
screen off/on happened, we somewhat get two extra sizeChanged callbacks. The first callback has
the message history view at full height, and the 2nd callback has it shrunk to accommodate the IME
keyboard height. In each callback, we call ListView.setSelection(), which seems to cause the
gratuitous scrolling jank.

I did two things to fix this:
1. in onPause(), remember the last scroll position for the Message history list view. After the
activity is resumed and the message list re-queried, I set the ListView's selection to the saved
scroll position (instead of always scroll to the end). This addressed Hiroshi's concern that
we shouldn't automatically scroll to the end after the screen turned back on.
2. in the 1st sizeChanged() callback, the Message history ListView's layout isn't completed yet.
So don't call scroll to end in that case.

These two things made the gratuitous scroll go away (well, because we are not scrolling to end
in the first place)

- fixed some logic in smooth scrolling. We don't need mFirstScrolled flag. It was used incorrectly
sometimes. Also, don't smooth scroll at all if the delta of the scroll is too big. That just seems
gratuitous.

-

Change-Id: I52a073061ac014089b78587bcda0ba64c97d3a78